+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added private function `Get-CommandParameters` to filter command parameters
+- Added private function `Get-CommandParameter` to filter command parameters
   by excluding specified parameter names and common parameters, providing a
   reusable way to determine settable properties on objects.